requirements.txt
Django==1.9.5
gunicorn==19.4.5
git+git://github.com/macropin/django-registration
django-crispy-forms-ng==2.0.0
python-magic==0.4.11
django-datatables-view==1.13.0
psycopg2==2.6.1
django-email-changer==0.1.2
django-storages-redux==1.3.2
boto==2.39.0
django-grappelli==2.8.1
djangorestframework==3.3.3
markdown==2.6.6
django-filter==0.13.0
PyVCF==0.6.8
bitarray==0.8.1
architect==0.5.3
django-extensions==1.6.1
pygraphviz==1.3.1
git+https://github.com/staphopia/scholar.py.git#egg=scholar
psutil==4.1.0
beautifulsoup4==4.4.1
lxml==3.6.0
elasticsearch==2.3.0
urllib3==1.15.1
django-haystack==2.4.1
